feat(user): restructure user components and update layouts for consistency
- Restructured user-related components into organized folders by feature (dashboard, FAQ, portfolio, profile, showcase, landing-page)
- Added new reusable components for dashboard, FAQ, portfolio, profile, and showcase views
- Updated layout files across super-admin and user pages to align with the new component structure
- Renamed landing page components and moved them under user scope
- Introduced type definition files for FAQ and profile components
- Added mock data file for portfolio items

This improves modularity, reusability, and maintainability of the user-facing features.
