+++ app/(super-admin)/dashboard-super-admin/layout.tsx
@@ -2,7 +2,7 @@
 import { ReactNode } from "react";
 
 import Navbar from "@/components/navbar/SuperAdminNavbar";
-import Footer from "@/components/landing-page/Footer";
+import Footer from "@/components/user/landing-page/Footer";
 
 interface DashboardLayoutProps {
   children: ReactNode;
