+++ app/(super-admin)/detail-super-admin/portfolio/view/layout.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
 import Navbar from "@/components/navbar/SuperAdminNavbar";
-import Footer from "@/components/landing-page/Footer";
+import Footer from "@/components/user/landing-page/Footer";
 
 interface DetailPortfolioLayoutViewProps {
   children: ReactNode;
