+++ app/(super-admin)/profile-super-admin/layout.tsx
@@ -1,7 +1,7 @@
 // profile/layout.tsx
 import { ReactNode } from "react";
 import Navbar from "@/components/navbar/SuperAdminNavbar";
-import Footer from "@/components/landing-page/Footer";
+import Footer from "@/components/user/landing-page/Footer";
 
 interface ProfileLayoutProps {
   children: ReactNode;
