+++ app/page.tsx
@@ -1,4 +1,4 @@
-import LandingPage from "@/components/landing-page/LandingPage";
+import LandingPage from "@/components/user/landing-page/LandingPage";
 
 export default function Home() {
   return <LandingPage />;
