+++ components/user/dashboard/DashboardHeader.tsx
@@ -1,0 +1,13 @@
+interface DashboardHeaderProps {
+  title: string;
+  description: string;
+}
+
+export function DashboardHeader({ title, description }: DashboardHeaderProps) {
+  return (
+    <div className="bg-white/5 backdrop-blur-sm rounded-xl p-8 mb-12">
+      <h1 className="text-3xl font-bold text-white mb-4">{title}</h1>
+      <p className="text-gray-300">{description}</p>
+    </div>
+  );
+}
