+++ components/user/landing-page/Hero.tsx
