+++ components/user/profile/ProfileImage.tsx
@@ -1,0 +1,55 @@
+import { User, Upload } from 'lucide-react';
+import Image from 'next/image';
+import { Button } from '@/components/ui/button';
+import { useRef } from 'react';
+
+interface ProfileImageProps {
+  profileImage: string;
+  onImageUpload: (e: React.ChangeEvent<HTMLInputElement>) => void;
+}
+
+export function ProfileImage({ profileImage, onImageUpload }: ProfileImageProps) {
+  const fileInputRef = useRef<HTMLInputElement>(null);
+
+  const triggerFileInput = () => {
+    fileInputRef.current?.click();
+  };
+
+  return (
+    <div className="relative w-48 h-48 rounded-full bg-white/10 overflow-hidden flex-shrink-0 group">
+      {profileImage === '/default-avatar.png' ? (
+        <div className="w-full h-full flex items-center justify-center bg-white/5">
+          <User className="w-24 h-24 text-white/50" />
+        </div>
+      ) : (
+        <div className="relative w-full h-full">
+          <Image
+            src={profileImage}
+            alt="Profile Picture"
+            fill
+            className="object-cover"
+            sizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
+            priority
+          />
+        </div>
+      )}
+      <div className="absolute inset-0 bg-black/50 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
+        <Button
+          variant="ghost"
+          size="icon"
+          onClick={triggerFileInput}
+          className="text-white hover:text-white hover:bg-blue-900"
+        >
+          <Upload className="h-6 w-6" />
+        </Button>
+      </div>
+      <input
+        type="file"
+        ref={fileInputRef}
+        onChange={onImageUpload}
+        accept="image/*"
+        className="hidden"
+      />
+    </div>
+  );
+}
