+++ components/user/showcase/ShowcaseDetail/ProjectSidebar.tsx
@@ -1,0 +1,47 @@
+import { LinkIcon, Users } from 'lucide-react';
+import type { PortfolioItem } from '../types';
+
+interface ProjectSidebarProps {
+  project: PortfolioItem;
+}
+
+export function ProjectSidebar({ project }: ProjectSidebarProps) {
+  return (
+    <div className="space-y-8">
+      {project.links && (
+        <div className="bg-white/5 backdrop-blur-sm rounded-xl p-6">
+          {project.links.map((link, index) => (
+            <div key={index}>
+              <h2 className="text-xl font-semibold text-white mb-2">
+                {link.title}
+              </h2>
+              <div className="flex items-center gap-2 bg-[#011B45] rounded-lg p-4 mb-6">
+                <LinkIcon className="h-6 w-6 text-blue-500" />
+                <a href={link.url} target="_blank" rel="noopener noreferrer" className="text-white hover:text-blue-400 transition-colors">
+                  {link.url}
+                </a>
+              </div>
+            </div>
+          ))}
+        </div>
+      )}
+
+      {project.teamMembers && (
+        <div className="bg-white/5 backdrop-blur-sm rounded-xl p-6">
+          <h2 className="text-white font-semibold mb-4 flex items-center gap-2">
+            <Users className="h-5 w-5" />
+            Team Project
+          </h2>
+          <div className="space-y-4">
+            {project.teamMembers.map((member, index) => (
+              <div key={index} className="text-gray-300">
+                <p className="text-white">{member.name}</p>
+                <p className="text-sm">{member.role}</p>
+              </div>
+            ))}
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
