+++ components/user/showcase/ShowcaseHeader.tsx
@@ -1,0 +1,12 @@
+export function ShowcaseHeader() {
+  return (
+    <div className="text-center mb-16">
+      <h1 className="text-4xl md:text-5xl font-bold text-white mb-6">
+        Portfolio Showcase
+      </h1>
+      <p className="text-gray-300 max-w-3xl mx-auto">
+        Lorem ipsum dolor sit amet consectetur. Quisque purus risus in purus at a. Tincidunt et sapien donec id integer pulvinar. Eu purus accumsan a ornare dictum massa mattis. Suspendisse at dolor
+      </p>
+    </div>
+  );
+}
