+++ data/portfolioItems.ts
@@ -1,0 +1,209 @@
+import type { PortfolioItem } from '@/components/user/showcase/types';
+
+export const portfolioItems: PortfolioItem[] = [
+  {
+    id: 1,
+    title: 'UI/UX Healthy Application',
+    image: '/portfolio-1.png',
+    category: 'UI/UX Design',
+    tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
+    date: 'Released November 21, 2023',
+    subtitle: 'Rekayasa Perangkat Lunak - 2022',
+    description: [
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam.',
+      'Porttitor molestie tincidunt neque turpis feugiat mattis. Diam arcu sed mi fusce vestibulum eu neque sed. Dis hendrerit ultrices ac sodales. Justo mauris elementum scelerisque convallis.',
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam. Porttitor molestie'
+    ],
+    links: [
+      {
+        title: 'Figma Mobile App',
+        url: 'http://www'
+      },
+      {
+        title: 'Medium Study Case',
+        url: 'http://www'
+      }
+    ],
+    teamMembers: [
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' }
+    ],
+    contact: {
+      name: 'Krisna Bimantoro',
+      id: '202210370311254'
+    }
+  },
+  // Duplicate item dengan id berbeda untuk contoh
+  {
+    id: 2,
+    title: 'UI/UX Healthy Application',
+    image: '/portfolio-1.png',
+    category: 'UI/UX Design',
+    tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
+    date: 'Released November 21, 2023',
+    subtitle: 'Rekayasa Perangkat Lunak - 2022',
+    description: [
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam.',
+      'Porttitor molestie tincidunt neque turpis feugiat mattis. Diam arcu sed mi fusce vestibulum eu neque sed. Dis hendrerit ultrices ac sodales. Justo mauris elementum scelerisque convallis.',
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam. Porttitor molestie'
+    ],
+    links: [
+      {
+        title: 'Figma Mobile App',
+        url: 'http://www'
+      },
+      {
+        title: 'Medium Study Case',
+        url: 'http://www'
+      }
+    ],
+    teamMembers: [
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' }
+    ],
+    contact: {
+      name: 'Krisna Bimantoro',
+      id: '202210370311254'
+    }
+  },
+  {
+    id: 3,
+    title: 'UI/UX Healthy Application',
+    image: '/portfolio-1.png',
+    category: 'UI/UX Design',
+    tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
+    date: 'Released November 21, 2023',
+    subtitle: 'Rekayasa Perangkat Lunak - 2022',
+    description: [
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam.',
+      'Porttitor molestie tincidunt neque turpis feugiat mattis. Diam arcu sed mi fusce vestibulum eu neque sed. Dis hendrerit ultrices ac sodales. Justo mauris elementum scelerisque convallis.',
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam. Porttitor molestie'
+    ],
+    links: [
+      {
+        title: 'Figma Mobile App',
+        url: 'http://www'
+      },
+      {
+        title: 'Medium Study Case',
+        url: 'http://www'
+      }
+    ],
+    teamMembers: [
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' }
+    ],
+    contact: {
+      name: 'Krisna Bimantoro',
+      id: '202210370311254'
+    }
+  },
+  {
+    id: 4,
+    title: 'UI/UX Healthy Application',
+    image: '/portfolio-1.png',
+    category: 'UI/UX Design',
+    tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
+    date: 'Released November 21, 2023',
+    subtitle: 'Rekayasa Perangkat Lunak - 2022',
+    description: [
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam.',
+      'Porttitor molestie tincidunt neque turpis feugiat mattis. Diam arcu sed mi fusce vestibulum eu neque sed. Dis hendrerit ultrices ac sodales. Justo mauris elementum scelerisque convallis.',
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam. Porttitor molestie'
+    ],
+    links: [
+      {
+        title: 'Figma Mobile App',
+        url: 'http://www'
+      },
+      {
+        title: 'Medium Study Case',
+        url: 'http://www'
+      }
+    ],
+    teamMembers: [
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' }
+    ],
+    contact: {
+      name: 'Krisna Bimantoro',
+      id: '202210370311254'
+    }
+  },
+  {
+    id: 5,
+    title: 'UI/UX Healthy Application',
+    image: '/portfolio-1.png',
+    category: 'UI/UX Design',
+    tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
+    date: 'Released November 21, 2023',
+    subtitle: 'Rekayasa Perangkat Lunak - 2022',
+    description: [
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam.',
+      'Porttitor molestie tincidunt neque turpis feugiat mattis. Diam arcu sed mi fusce vestibulum eu neque sed. Dis hendrerit ultrices ac sodales. Justo mauris elementum scelerisque convallis.',
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam. Porttitor molestie'
+    ],
+    links: [
+      {
+        title: 'Figma Mobile App',
+        url: 'http://www'
+      },
+      {
+        title: 'Medium Study Case',
+        url: 'http://www'
+      }
+    ],
+    teamMembers: [
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' }
+    ],
+    contact: {
+      name: 'Krisna Bimantoro',
+      id: '202210370311254'
+    }
+  },
+  {
+    id: 6,
+    title: 'UI/UX Healthy Application',
+    image: '/portfolio-1.png',
+    category: 'UI/UX Design',
+    tags: ['Mobile Application', 'UI/UX Design', 'Design System'],
+    date: 'Released November 21, 2023',
+    subtitle: 'Rekayasa Perangkat Lunak - 2022',
+    description: [
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam.',
+      'Porttitor molestie tincidunt neque turpis feugiat mattis. Diam arcu sed mi fusce vestibulum eu neque sed. Dis hendrerit ultrices ac sodales. Justo mauris elementum scelerisque convallis.',
+      'Lorem ipsum dolor sit amet consectetur. In non iaculis scelerisque feugiat urna. Neque curabitur ac pretium ut elementum sed vel lobortis. Dui turpis nisl facilisis in vitae consectetur accumsan in. Sed tristique est sit semper. Scelerisque dui nunc mi gravida dignissim erat ipsum consequat aliquam. Porttitor molestie'
+    ],
+    links: [
+      {
+        title: 'Figma Mobile App',
+        url: 'http://www'
+      },
+      {
+        title: 'Medium Study Case',
+        url: 'http://www'
+      }
+    ],
+    teamMembers: [
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' },
+      { name: 'Krisna Bimantoro', role: 'UI/UX Designer' }
+    ],
+    contact: {
+      name: 'Krisna Bimantoro',
+      id: '202210370311254'
+    }
+  }
+];
